all environment queries are moved from resolvers and services to repositories
