+++ backend/dataall/core/environment/db/environment_repositories.py
@@ -1,5 +1,11 @@
-from dataall.core.environment.db.environment_models import EnvironmentParameter, Environment
-from sqlalchemy.sql import and_
+from dataall.core.environment.db.environment_models import (
+    EnvironmentParameter,
+    Environment,
+    ConsumptionRole,
+    EnvironmentGroup,
+)
+from sqlalchemy.sql import and_, or_
+from sqlalchemy.orm import Query
 
 from dataall.base.db import exceptions
 
@@ -67,3 +73,231 @@ class EnvironmentRepository:
         if not environment:
             return None
         return environment
+
+    @staticmethod
+    def get_environment_consumption_role(session, role_uri, environment_uri) -> ConsumptionRole:
+        return (
+            session.query(ConsumptionRole)
+            .filter(
+                (
+                    and_(
+                        ConsumptionRole.consumptionRoleUri == role_uri,
+                        ConsumptionRole.environmentUri == environment_uri,
+                    )
+                )
+            )
+            .first()
+        )
+
+    @staticmethod
+    def get_environment_group(session, group_uri, environment_uri):
+        return (
+            session.query(EnvironmentGroup)
+            .filter(
+                (
+                    and_(
+                        EnvironmentGroup.groupUri == group_uri,
+                        EnvironmentGroup.environmentUri == environment_uri,
+                    )
+                )
+            )
+            .first()
+        )
+
+    @staticmethod
+    def get_consumption_role(session, uri):
+        return (
+            session.query(ConsumptionRole)
+            .filter(
+                and_(
+                    ConsumptionRole.consumptionRoleUri == uri,
+                )
+            )
+            .first()
+        )
+
+    @staticmethod
+    def find_consumption_roles_by_IAMArn(session, uri, arn):
+        return (
+            session.query(ConsumptionRole)
+            .filter(and_(ConsumptionRole.environmentUri == uri, ConsumptionRole.IAMRoleArn == arn))
+            .first()
+        )
+
+    @staticmethod
+    def query_all_environment_consumption_roles(session, uri, filter) -> Query:
+        query = session.query(ConsumptionRole).filter(ConsumptionRole.environmentUri == uri)
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    ConsumptionRole.consumptionRoleName.ilike('%' + term + '%'),
+                )
+            )
+        if filter and filter.get('groupUri'):
+            group = filter['groupUri']
+            query = query.filter(
+                or_(
+                    ConsumptionRole.groupUri == group,
+                )
+            )
+        return query.order_by(ConsumptionRole.consumptionRoleUri)
+
+    @staticmethod
+    def query_user_environment_consumption_roles(session, groups, uri, filter) -> Query:
+        query = (
+            session.query(ConsumptionRole)
+            .filter(ConsumptionRole.environmentUri == uri)
+            .filter(ConsumptionRole.groupUri.in_(groups))
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    ConsumptionRole.consumptionRoleName.ilike('%' + term + '%'),
+                )
+            )
+        if filter and filter.get('groupUri'):
+            print('filter group')
+            group = filter['groupUri']
+            query = query.filter(
+                or_(
+                    ConsumptionRole.groupUri == group,
+                )
+            )
+        return query.order_by(ConsumptionRole.consumptionRoleUri)
+
+    @staticmethod
+    def query_environment_invited_groups(session, uri, filter) -> Query:
+        query = (
+            session.query(EnvironmentGroup)
+            .join(
+                Environment,
+                EnvironmentGroup.environmentUri == Environment.environmentUri,
+            )
+            .filter(
+                and_(
+                    Environment.environmentUri == uri,
+                    EnvironmentGroup.groupUri != Environment.SamlGroupName,
+                )
+            )
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
+                )
+            )
+        return query
+
+    @staticmethod
+    def query_user_environment_groups(session, groups, uri, filter) -> Query:
+        query = (
+            session.query(EnvironmentGroup)
+            .filter(EnvironmentGroup.environmentUri == uri)
+            .filter(EnvironmentGroup.groupUri.in_(groups))
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
+                )
+            )
+        return query
+
+    @staticmethod
+    def query_all_environment_groups(session, uri, filter) -> Query:
+        query = session.query(EnvironmentGroup).filter(EnvironmentGroup.environmentUri == uri)
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
+                )
+            )
+        return query
+
+    @staticmethod
+    def query_user_consumption_roles(session, username, groups, filter) -> Query:
+        query = (
+            session.query(ConsumptionRole)
+            .filter(ConsumptionRole.groupUri.in_(groups))
+            .distinct(ConsumptionRole.consumptionRoleName)
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    ConsumptionRole.consumptionRoleName.ilike('%' + term + '%'),
+                )
+            )
+        if filter and filter.get('groupUri'):
+            print('filter group')
+            group = filter['groupUri']
+            query = query.filter(
+                or_(
+                    ConsumptionRole.groupUri == group,
+                )
+            )
+        return query
+
+    @staticmethod
+    def query_user_groups(session, username, groups, filter) -> Query:
+        query = (
+            session.query(EnvironmentGroup)
+            .filter(EnvironmentGroup.groupUri.in_(groups))
+            .distinct(EnvironmentGroup.groupUri)
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
+                )
+            )
+        return query
+
+    @staticmethod
+    def query_user_environments(session, username, groups, filter) -> Query:
+        query = (
+            session.query(Environment)
+            .outerjoin(
+                EnvironmentGroup,
+                Environment.environmentUri == EnvironmentGroup.environmentUri,
+            )
+            .filter(
+                or_(
+                    Environment.owner == username,
+                    EnvironmentGroup.groupUri.in_(groups),
+                )
+            )
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    Environment.label.ilike('%' + term + '%'),
+                    Environment.description.ilike('%' + term + '%'),
+                    Environment.tags.contains(f'{{{term}}}'),
+                    Environment.region.ilike('%' + term + '%'),
+                )
+            )
+        if filter and filter.get('SamlGroupName') and filter.get('SamlGroupName') in groups:
+            query = query.filter(EnvironmentGroup.groupUri == filter.get('SamlGroupName'))
+        return query
+
+    @staticmethod
+    def is_user_invited_to_environment(session, groups, uri):
+        env_group = (
+            session.query(EnvironmentGroup)
+            .filter(
+                and_(
+                    EnvironmentGroup.environmentUri == uri,
+                    EnvironmentGroup.groupUri.in_(groups),
+                )
+            )
+            .first()
+        )
+        return env_group is not None
