+++ backend/dataall/core/environment/services/environment_service.py
@@ -1,9 +1,7 @@
 import logging
 import re
 
-from sqlalchemy import or_
 from sqlalchemy.orm import Query
-from sqlalchemy.sql import and_
 
 from dataall.base.context import get_context
 from dataall.core.permissions.services.resource_policy_service import ResourcePolicyService
@@ -16,7 +14,6 @@ from dataall.core.environment.services.environment_resource_manager import Envir
 from dataall.core.permissions.db.permission.permission_repositories import PermissionRepository
 from dataall.core.permissions.db.permission.permission_models import PermissionType
 
-from dataall.core.vpc.db.vpc_models import Vpc
 from dataall.base.db.paginator import paginate
 from dataall.base.utils.naming_convention import (
     NamingConventionService,
@@ -35,17 +32,101 @@ from dataall.core.environment.services.managed_iam_policies import PolicyManager
 from dataall.core.permissions.services.organization_permissions import LINK_ENVIRONMENT
 from dataall.core.permissions.services import environment_permissions
 from dataall.core.permissions.services.tenant_permissions import MANAGE_ENVIRONMENTS
+from dataall.core.vpc.db.vpc_repositories import VpcRepository
 
 log = logging.getLogger(__name__)
 
 
+class EnvironmentRequestValidationService:
+    @staticmethod
+    def validate_update_consumption_role(data):
+        if not data:
+            raise exceptions.RequiredParameter('input')
+        if not data.get('groupUri'):
+            raise exceptions.RequiredParameter('groupUri')
+        if not data.get('consumptionRoleName'):
+            raise exceptions.RequiredParameter('consumptionRoleName')
+
+    @staticmethod
+    def validate_invite_params(data):
+        if not data:
+            raise exceptions.RequiredParameter('data')
+        if not data.get('groupUri'):
+            raise exceptions.RequiredParameter('groupUri')
+        if not data.get('permissions'):
+            raise exceptions.RequiredParameter('permissions')
+
+    @staticmethod
+    def validate_creation_params(data, uri, session):
+        if not uri:
+            raise exceptions.RequiredParameter('organizationUri')
+        if not data:
+            raise exceptions.RequiredParameter('data')
+        if not data.get('label'):
+            raise exceptions.RequiredParameter('label')
+        if not data.get('SamlGroupName'):
+            raise exceptions.RequiredParameter('group')
+        if not data.get('AwsAccountId'):
+            raise exceptions.RequiredParameter('AwsAccountId')
+        if not data.get('region'):
+            raise exceptions.RequiredParameter('region')
+        EnvironmentRequestValidationService.validate_resource_prefix(data)
+        EnvironmentRequestValidationService.validate_account_region(data, session)
+
+    @staticmethod
+    def validate_resource_prefix(data):
+        if data.get('resourcePrefix') and not bool(re.match(r'^[a-z-]+$', data.get('resourcePrefix'))):
+            raise exceptions.InvalidInput(
+                'resourcePrefix',
+                data.get('resourcePrefix'),
+                'must match the pattern ^[a-z-]+$',
+            )
+
+    @staticmethod
+    def validate_account_region(data, session):
+        environment = EnvironmentRepository.find_environment_by_account_region(
+            session=session, account_id=data.get('AwsAccountId'), region=data.get('region')
+        )
+        if environment:
+            raise exceptions.InvalidInput(
+                'AwsAccount/region',
+                f"{data.get('AwsAccountId')}/{data.get('region')}",
+                f"unique. An environment for {data.get('AwsAccountId')}/{data.get('region')} already exists",
+            )
+
+
 class EnvironmentService:
+    @staticmethod
+    def validate_permissions(session, uri, g_permissions, group):
+        """
+        g_permissions: coming from frontend = ENVIRONMENT_INVITATION_REQUEST
+
+        """
+        if environment_permissions.INVITE_ENVIRONMENT_GROUP in g_permissions:
+            g_permissions.append(environment_permissions.REMOVE_ENVIRONMENT_GROUP)
+
+        g_permissions.extend(environment_permissions.ENVIRONMENT_INVITED_DEFAULT)
+        g_permissions = list(set(g_permissions))
+
+        if g_permissions not in environment_permissions.ENVIRONMENT_INVITED:
+            exceptions.PermissionUnauthorized(action='INVITE_TEAM', group_name=group, resource_uri=uri)
+
+        env_group_permissions = []
+        for p in g_permissions:
+            env_group_permissions.append(
+                PermissionRepository.find_permission_by_name(
+                    session=session,
+                    permission_name=p,
+                    permission_type=PermissionType.RESOURCE.name,
+                )
+            )
+
     @staticmethod
     @TenantPolicyService.has_tenant_permission(MANAGE_ENVIRONMENTS)
     @ResourcePolicyService.has_resource_permission(LINK_ENVIRONMENT)
     def create_environment(session, uri, data=None):
         context = get_context()
-        EnvironmentService._validate_creation_params(data, uri, session)
+        EnvironmentRequestValidationService.validate_creation_params(data, uri, session)
         organization = OrganizationRepository.get_organization_by_uri(session, uri)
         env = Environment(
             organizationUri=data.get('organizationUri'),
@@ -130,48 +211,11 @@ class EnvironmentService:
         return env
 
     @staticmethod
-    def _validate_creation_params(data, uri, session):
-        if not uri:
-            raise exceptions.RequiredParameter('organizationUri')
-        if not data:
-            raise exceptions.RequiredParameter('data')
-        if not data.get('label'):
-            raise exceptions.RequiredParameter('label')
-        if not data.get('SamlGroupName'):
-            raise exceptions.RequiredParameter('group')
-        if not data.get('AwsAccountId'):
-            raise exceptions.RequiredParameter('AwsAccountId')
-        if not data.get('region'):
-            raise exceptions.RequiredParameter('region')
-        EnvironmentService._validate_resource_prefix(data)
-        EnvironmentService._validate_account_region(data, session)
-
-    @staticmethod
-    def _validate_resource_prefix(data):
-        if data.get('resourcePrefix') and not bool(re.match(r'^[a-z-]+$', data.get('resourcePrefix'))):
-            raise exceptions.InvalidInput(
-                'resourcePrefix',
-                data.get('resourcePrefix'),
-                'must match the pattern ^[a-z-]+$',
-            )
-
-    @staticmethod
-    def _validate_account_region(data, session):
-        environment = EnvironmentRepository.find_environment_by_account_region(
-            session=session, account_id=data.get('AwsAccountId'), region=data.get('region')
-        )
-        if environment:
-            raise exceptions.InvalidInput(
-                'AwsAccount/region',
-                f"{data.get('AwsAccountId')}/{data.get('region')}",
-                f"unique. An environment for {data.get('AwsAccountId')}/{data.get('region')} already exists",
-            )
-
     @staticmethod
     @TenantPolicyService.has_tenant_permission(MANAGE_ENVIRONMENTS)
     @ResourcePolicyService.has_resource_permission(environment_permissions.UPDATE_ENVIRONMENT)
     def update_environment(session, uri, data=None):
-        EnvironmentService._validate_resource_prefix(data)
+        EnvironmentRequestValidationService.validate_resource_prefix(data)
         environment = EnvironmentService.get_environment_by_uri(session, uri)
         if data.get('label'):
             environment.label = data.get('label')
@@ -208,7 +252,7 @@ class EnvironmentService:
     @TenantPolicyService.has_tenant_permission(MANAGE_ENVIRONMENTS)
     @ResourcePolicyService.has_resource_permission(environment_permissions.INVITE_ENVIRONMENT_GROUP)
     def invite_group(session, uri, data=None) -> (Environment, EnvironmentGroup):
-        EnvironmentService.validate_invite_params(data)
+        EnvironmentRequestValidationService.validate_invite_params(data)
 
         group: str = data['groupUri']
 
@@ -275,31 +319,6 @@ class EnvironmentService:
         return environment, environment_group
 
     @staticmethod
-    def validate_permissions(session, uri, g_permissions, group):
-        """
-        g_permissions: coming from frontend = ENVIRONMENT_INVITATION_REQUEST
-
-        """
-        if environment_permissions.INVITE_ENVIRONMENT_GROUP in g_permissions:
-            g_permissions.append(environment_permissions.REMOVE_ENVIRONMENT_GROUP)
-
-        g_permissions.extend(environment_permissions.ENVIRONMENT_INVITED_DEFAULT)
-        g_permissions = list(set(g_permissions))
-
-        if g_permissions not in environment_permissions.ENVIRONMENT_INVITED:
-            exceptions.PermissionUnauthorized(action='INVITE_TEAM', group_name=group, resource_uri=uri)
-
-        env_group_permissions = []
-        for p in g_permissions:
-            env_group_permissions.append(
-                PermissionRepository.find_permission_by_name(
-                    session=session,
-                    permission_name=p,
-                    permission_type=PermissionType.RESOURCE.name,
-                )
-            )
-
-    @staticmethod
     @TenantPolicyService.has_tenant_permission(MANAGE_ENVIRONMENTS)
     @ResourcePolicyService.has_resource_permission(environment_permissions.REMOVE_ENVIRONMENT_GROUP)
     def remove_group(session, uri, group):
@@ -321,7 +340,7 @@ class EnvironmentService:
                 message=f'Team: {group} has created {group_env_objects_count} resources on this environment.',
             )
 
-        group_env_consumption_roles = EnvironmentService.query_user_environment_consumption_roles(
+        group_env_consumption_roles = EnvironmentRepository.query_user_environment_consumption_roles(
             session, [group], uri, {}
         ).all()
         if group_env_consumption_roles:
@@ -356,7 +375,7 @@ class EnvironmentService:
     @TenantPolicyService.has_tenant_permission(MANAGE_ENVIRONMENTS)
     @ResourcePolicyService.has_resource_permission(environment_permissions.UPDATE_ENVIRONMENT_GROUP)
     def update_group_permissions(session, uri, data=None):
-        EnvironmentService.validate_invite_params(data)
+        EnvironmentRequestValidationService.validate_invite_params(data)
 
         group = data['groupUri']
 
@@ -424,7 +443,7 @@ class EnvironmentService:
         IAMRoleArn: str = data['IAMRoleArn']
         environment = EnvironmentService.get_environment_by_uri(session, uri)
 
-        alreadyAdded = EnvironmentService.find_consumption_roles_by_IAMArn(
+        alreadyAdded = EnvironmentRepository.find_consumption_roles_by_IAMArn(
             session, environment.environmentUri, IAMRoleArn
         )
         if alreadyAdded:
@@ -501,12 +520,7 @@ class EnvironmentService:
     @TenantPolicyService.has_tenant_permission(MANAGE_ENVIRONMENTS)
     @ResourcePolicyService.has_resource_permission(environment_permissions.REMOVE_ENVIRONMENT_CONSUMPTION_ROLE)
     def update_consumption_role(session, uri, env_uri, input):
-        if not input:
-            raise exceptions.RequiredParameter('input')
-        if not input.get('groupUri'):
-            raise exceptions.RequiredParameter('groupUri')
-        if not input.get('consumptionRoleName'):
-            raise exceptions.RequiredParameter('consumptionRoleName')
+        EnvironmentRequestValidationService.validate_update_consumption_role(input)
         consumption_role = EnvironmentService.get_environment_consumption_role(session, uri, env_uri)
         if consumption_role:
             ResourcePolicyService.update_resource_policy(
@@ -523,39 +537,10 @@ class EnvironmentService:
         return consumption_role
 
     @staticmethod
-    def query_user_environments(session, username, groups, filter) -> Query:
-        query = (
-            session.query(Environment)
-            .outerjoin(
-                EnvironmentGroup,
-                Environment.environmentUri == EnvironmentGroup.environmentUri,
-            )
-            .filter(
-                or_(
-                    Environment.owner == username,
-                    EnvironmentGroup.groupUri.in_(groups),
-                )
-            )
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    Environment.label.ilike('%' + term + '%'),
-                    Environment.description.ilike('%' + term + '%'),
-                    Environment.tags.contains(f'{{{term}}}'),
-                    Environment.region.ilike('%' + term + '%'),
-                )
-            )
-        if filter and filter.get('SamlGroupName') and filter.get('SamlGroupName') in groups:
-            query = query.filter(EnvironmentGroup.groupUri == filter.get('SamlGroupName'))
-        return query
-
-    @staticmethod
     def paginated_user_environments(session, data=None) -> dict:
         context = get_context()
         return paginate(
-            query=EnvironmentService.query_user_environments(session, context.username, context.groups, data),
+            query=EnvironmentRepository.query_user_environments(session, context.username, context.groups, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 5),
         ).to_dict()
@@ -563,7 +548,7 @@ class EnvironmentService:
     @staticmethod
     def list_valid_user_environments(session, data=None) -> dict:
         context = get_context()
-        query = EnvironmentService.query_user_environments(session, context.username, context.groups, data)
+        query = EnvironmentRepository.query_user_environments(session, context.username, context.groups, data)
         valid_environments = []
         for env in query:
             stack = stack_helper.get_stack_with_cfn_resources(
@@ -583,105 +568,41 @@ class EnvironmentService:
         }
 
     @staticmethod
-    def query_user_groups(session, username, groups, filter) -> Query:
-        query = (
-            session.query(EnvironmentGroup)
-            .filter(EnvironmentGroup.groupUri.in_(groups))
-            .distinct(EnvironmentGroup.groupUri)
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
-                )
-            )
-        return query
-
-    @staticmethod
     def paginated_user_groups(session, data=None) -> dict:
         context = get_context()
         return paginate(
-            query=EnvironmentService.query_user_groups(session, context.username, context.groups, data),
+            query=EnvironmentRepository.query_user_groups(session, context.username, context.groups, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 5),
         ).to_dict()
-
-    @staticmethod
-    def query_user_consumption_roles(session, username, groups, filter) -> Query:
-        query = (
-            session.query(ConsumptionRole)
-            .filter(ConsumptionRole.groupUri.in_(groups))
-            .distinct(ConsumptionRole.consumptionRoleName)
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    ConsumptionRole.consumptionRoleName.ilike('%' + term + '%'),
-                )
-            )
-        if filter and filter.get('groupUri'):
-            print('filter group')
-            group = filter['groupUri']
-            query = query.filter(
-                or_(
-                    ConsumptionRole.groupUri == group,
-                )
-            )
-        return query
 
     @staticmethod
     def paginated_user_consumption_roles(session, data=None) -> dict:
         context = get_context()
         return paginate(
-            query=EnvironmentService.query_user_consumption_roles(session, context.username, context.groups, data),
+            query=EnvironmentRepository.query_user_consumption_roles(session, context.username, context.groups, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 5),
         ).to_dict()
 
     @staticmethod
-    def query_user_environment_groups(session, groups, uri, filter) -> Query:
-        query = (
-            session.query(EnvironmentGroup)
-            .filter(EnvironmentGroup.environmentUri == uri)
-            .filter(EnvironmentGroup.groupUri.in_(groups))
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
-                )
-            )
-        return query
-
-    @staticmethod
     @ResourcePolicyService.has_resource_permission(environment_permissions.LIST_ENVIRONMENT_GROUPS)
     def paginated_user_environment_groups(session, uri, data=None) -> dict:
         return paginate(
-            query=EnvironmentService.query_user_environment_groups(session, get_context().groups, uri, data),
+            query=EnvironmentRepository.query_user_environment_groups(session, get_context().groups, uri, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 1000),
         ).to_dict()
 
     @staticmethod
-    def query_all_environment_groups(session, uri, filter) -> Query:
-        query = session.query(EnvironmentGroup).filter(EnvironmentGroup.environmentUri == uri)
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
-                )
-            )
-        return query
+    def get_all_environment_groups(session, uri, filter) -> Query:
+        return EnvironmentRepository.query_all_environment_groups(session, uri, filter)
 
     @staticmethod
     @ResourcePolicyService.has_resource_permission(environment_permissions.LIST_ENVIRONMENT_GROUPS)
     def paginated_all_environment_groups(session, uri, data=None) -> dict:
         return paginate(
-            query=EnvironmentService.query_all_environment_groups(session, uri, data),
+            query=EnvironmentService.get_all_environment_groups(session, uri, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 10),
         ).to_dict()
@@ -691,159 +612,54 @@ class EnvironmentService:
     def list_environment_groups(session, uri) -> [str]:
         return [
             g.groupUri
-            for g in EnvironmentService.query_user_environment_groups(session, get_context().groups, uri, {}).all()
+            for g in EnvironmentRepository.query_user_environment_groups(session, get_context().groups, uri, {}).all()
         ]
-
-    @staticmethod
-    def query_environment_invited_groups(session, uri, filter) -> Query:
-        query = (
-            session.query(EnvironmentGroup)
-            .join(
-                Environment,
-                EnvironmentGroup.environmentUri == Environment.environmentUri,
-            )
-            .filter(
-                and_(
-                    Environment.environmentUri == uri,
-                    EnvironmentGroup.groupUri != Environment.SamlGroupName,
-                )
-            )
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    EnvironmentGroup.groupUri.ilike('%' + term + '%'),
-                )
-            )
-        return query
 
     @staticmethod
     @ResourcePolicyService.has_resource_permission(environment_permissions.LIST_ENVIRONMENT_GROUPS)
     def paginated_environment_invited_groups(session, uri, data=None) -> dict:
         return paginate(
-            query=EnvironmentService.query_environment_invited_groups(session, uri, data),
+            query=EnvironmentRepository.query_environment_invited_groups(session, uri, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 10),
         ).to_dict()
 
     @staticmethod
     def list_environment_invited_groups(session, uri):
-        return EnvironmentService.query_environment_invited_groups(session, uri, {}).all()
-
-    @staticmethod
-    def query_user_environment_consumption_roles(session, groups, uri, filter) -> Query:
-        query = (
-            session.query(ConsumptionRole)
-            .filter(ConsumptionRole.environmentUri == uri)
-            .filter(ConsumptionRole.groupUri.in_(groups))
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    ConsumptionRole.consumptionRoleName.ilike('%' + term + '%'),
-                )
-            )
-        if filter and filter.get('groupUri'):
-            print('filter group')
-            group = filter['groupUri']
-            query = query.filter(
-                or_(
-                    ConsumptionRole.groupUri == group,
-                )
-            )
-        return query.order_by(ConsumptionRole.consumptionRoleUri)
+        return EnvironmentRepository.query_environment_invited_groups(session, uri, {}).all()
 
     @staticmethod
     @ResourcePolicyService.has_resource_permission(environment_permissions.LIST_ENVIRONMENT_CONSUMPTION_ROLES)
     def paginated_user_environment_consumption_roles(session, uri, data=None) -> dict:
         return paginate(
-            query=EnvironmentService.query_user_environment_consumption_roles(session, get_context().groups, uri, data),
+            query=EnvironmentRepository.query_user_environment_consumption_roles(
+                session, get_context().groups, uri, data
+            ),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 1000),
         ).to_dict()
 
     @staticmethod
-    def query_all_environment_consumption_roles(session, uri, filter) -> Query:
-        query = session.query(ConsumptionRole).filter(ConsumptionRole.environmentUri == uri)
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    ConsumptionRole.consumptionRoleName.ilike('%' + term + '%'),
-                )
-            )
-        if filter and filter.get('groupUri'):
-            group = filter['groupUri']
-            query = query.filter(
-                or_(
-                    ConsumptionRole.groupUri == group,
-                )
-            )
-        return query.order_by(ConsumptionRole.consumptionRoleUri)
-
-    @staticmethod
     @ResourcePolicyService.has_resource_permission(environment_permissions.LIST_ENVIRONMENT_CONSUMPTION_ROLES)
     def paginated_all_environment_consumption_roles(session, uri, data=None) -> dict:
         return paginate(
-            query=EnvironmentService.query_all_environment_consumption_roles(session, uri, data),
+            query=EnvironmentRepository.query_all_environment_consumption_roles(session, uri, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 10),
         ).to_dict()
 
     @staticmethod
-    def find_consumption_roles_by_IAMArn(session, uri, arn) -> Query:
-        return (
-            session.query(ConsumptionRole)
-            .filter(and_(ConsumptionRole.environmentUri == uri, ConsumptionRole.IAMRoleArn == arn))
-            .first()
-        )
-
-    @staticmethod
     def get_consumption_role(session, uri) -> Query:
-        return (
-            session.query(ConsumptionRole)
-            .filter(
-                and_(
-                    ConsumptionRole.consumptionRoleUri == uri,
-                )
-            )
-            .first()
-        )
-
-    @staticmethod
-    def query_environment_networks(session, uri, filter) -> Query:
-        query = session.query(Vpc).filter(
-            Vpc.environmentUri == uri,
-        )
-        if filter.get('term'):
-            term = filter.get('term')
-            query = query.filter(
-                or_(
-                    Vpc.label.ilike('%' + term + '%'),
-                    Vpc.VpcId.ilike('%' + term + '%'),
-                )
-            )
-        return query
+        return EnvironmentRepository.get_consumption_role(session, uri)
 
     @staticmethod
     @ResourcePolicyService.has_resource_permission(environment_permissions.LIST_ENVIRONMENT_NETWORKS)
     def paginated_environment_networks(session, uri, data=None) -> dict:
         return paginate(
-            query=EnvironmentService.query_environment_networks(session, uri, data),
+            query=VpcRepository.query_environment_networks(session, uri, data),
             page=data.get('page', 1),
             page_size=data.get('pageSize', 10),
         ).to_dict()
-
-    @staticmethod
-    def validate_invite_params(data):
-        if not data:
-            raise exceptions.RequiredParameter('data')
-        if not data.get('groupUri'):
-            raise exceptions.RequiredParameter('groupUri')
-        if not data.get('permissions'):
-            raise exceptions.RequiredParameter('permissions')
 
     @staticmethod
     def find_environment_group(session, group_uri, environment_uri):
@@ -855,36 +671,14 @@ class EnvironmentService:
 
     @staticmethod
     def get_environment_group(session, group_uri, environment_uri):
-        env_group = (
-            session.query(EnvironmentGroup)
-            .filter(
-                (
-                    and_(
-                        EnvironmentGroup.groupUri == group_uri,
-                        EnvironmentGroup.environmentUri == environment_uri,
-                    )
-                )
-            )
-            .first()
-        )
+        env_group = EnvironmentRepository.get_environment_group(session, group_uri, environment_uri)
         if not env_group:
             raise exceptions.ObjectNotFound('EnvironmentGroup', f'({group_uri},{environment_uri})')
         return env_group
 
     @staticmethod
     def get_environment_consumption_role(session, role_uri, environment_uri) -> ConsumptionRole:
-        role = (
-            session.query(ConsumptionRole)
-            .filter(
-                (
-                    and_(
-                        ConsumptionRole.consumptionRoleUri == role_uri,
-                        ConsumptionRole.environmentUri == environment_uri,
-                    )
-                )
-            )
-            .first()
-        )
+        role = EnvironmentRepository.get_environment_consumption_role(session, role_uri, environment_uri)
         if not role:
             raise exceptions.ObjectNotFound('ConsumptionRoleUri', f'({role_uri},{environment_uri})')
         return role
@@ -969,3 +763,9 @@ class EnvironmentService:
     def get_boolean_env_param(session, env: Environment, param: str) -> bool:
         param = EnvironmentParameterRepository(session).get_param(env.environmentUri, param)
         return param is not None and param.value.lower() == 'true'
+
+    @staticmethod
+    def is_user_invited(uri):
+        context = get_context()
+        with context.db_engine.scoped_session() as session:
+            return EnvironmentRepository.is_user_invited_to_environment(session=session, groups=context.groups, uri=uri)
